Analyzer/requirements.txt
networkx
pyyaml
flask
prometheus_client
requests